backend/requirements.txt
django
djangorestframework
google-generativeai
django-cors-headers
python-dotenv
PyPDF2
